+++ capycli/project/project_component_check.py
@@ -10,9 +10,8 @@ import logging
 import sys
 from typing import Any
 
-import sw360
-
 import capycli.common.script_base
+import sw360
 from capycli.bom.component_check import ComponentCheck
 from capycli.common.print import print_red, print_text, print_yellow
 from capycli.main.result_codes import ResultCode
@@ -26,13 +25,17 @@ class ProjectComponentCheck(capycli.common.script_base.ScriptBase):
     """
     def __init__(self) -> None:
         self.component_check = ComponentCheck()
+        self.verbose = False
 
     def is_dev_dependency(self, name: str) -> bool:
         """Check whether the given component matches any known development dependency."""
         dd = self.component_check.component_check_list.get("dev_dependencies", [])
         for ecosystem in dd:
             for entry in dd.get(ecosystem, []):
-                if name.lower() == entry.get("name", ""):
+                to_compare = entry.get("name", "")
+                if entry.get("namespace", ""):
+                    to_compare = entry.get("namespace", "") + "/" + to_compare
+                if name.lower() == to_compare.lower():
                     return True
 
         return False
@@ -129,11 +132,14 @@ class ProjectComponentCheck(capycli.common.script_base.ScriptBase):
             print_text("    --forceerror          force an error exit code in case of validation errors or warnings")
             return
 
-        print_text("Reading component check list from component_checks.json...")
+        self.verbose = args.verbose
+        self.component_check.verbose = args.verbose
+
+        print_text("Reading component checklist...")
         try:
             self.component_check.read_component_check_list(args.remote_check_list, args.local_checklist_list)
         except Exception as ex:
-            print_red("Error reading component check list " + repr(ex))
+            print_red("Error reading component checklist " + repr(ex))
             sys.exit(ResultCode.RESULT_GENERAL_ERROR)
         if len(self.component_check.component_check_list) > 0:
             print_text("  Got component checklist.")